Original Files
Finished up the two machines Alex didn't get too. Add a .gitignore file to less the size of the repo.

Later, I'll be adding SSH keys to the systems so that changes are tracked.

When I'm finished, feel free to edit the source code through either the VMs or locally.

I'll be working on getting instructions on how to run the files locally sometime soon.